requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.1
antlr4-python3-runtime==4.8
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autoalbument==0.4.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
efficientnet==1.0.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==0.10.0
flatbuffers==1.12
fonttools==4.37.0
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
hydra-core==1.0.6
idna==3.3
image-classifiers==1.0.0
imageio==2.21.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
munch==2.5.0
nest-asyncio==1.5.5
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.0.6
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.1.8
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
segmentation-models==1.0.1
segmentation-models-pytorch==0.3.0
six==1.16.0
stack-data==0.4.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.4.12
torch==1.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
